Renamed ThriftClientType for scanning a tablet, moved halt and fastHalt back to TabletClientService
