+++ server/tserver/src/main/java/org/apache/accumulo/tserver/TabletClientHandler.java
@@ -1528,4 +1528,32 @@ public class TabletClientHandler implements TabletClientService.Iface {
       return handleTimeout(sessionId);
     }
   }
+
+  @Override
+  public void halt(TInfo tinfo, TCredentials credentials, String lock)
+      throws ThriftSecurityException {
+
+    checkPermission(credentials, lock, "halt");
+
+    Halt.halt(0, () -> {
+      log.info("Manager requested tablet server halt");
+      server.gcLogger.logGCInfo(server.getConfiguration());
+      server.requestStop();
+      try {
+        server.getLock().unlock();
+      } catch (Exception e) {
+        log.error("Caught exception unlocking TabletServer lock", e);
+      }
+    });
+  }
+
+  @Override
+  public void fastHalt(TInfo info, TCredentials credentials, String lock) {
+    try {
+      halt(info, credentials, lock);
+    } catch (Exception e) {
+      log.warn("Error halting", e);
+    }
+  }
+
 }
